requirements.txt
asgiref==3.11.0
crispy-bootstrap5==2025.6
Django==5.2.8
django-crispy-forms==2.5
mysqlclient==2.2.7
sqlparse==0.5.3
tzdata==2025.2
